+++ baybe/core.py
@@ -13,7 +13,7 @@ from pydantic import Field, parse_obj_as
 from baybe.constraints import Constraint
 from baybe.parameters import Parameter
 from baybe.searchspace import SearchSpace
-from baybe.strategy import Strategy
+from baybe.strategies.strategy import Strategy
 from baybe.targets import Objective, Target
 from baybe.utils import BaseModel
 
@@ -23,19 +23,21 @@ log = logging.getLogger(__name__)
 class BayBE(BaseModel):
     """Main class for interaction with BayBE."""
 
+    # DOE specifications
+    searchspace: SearchSpace
     objective: Objective
     strategy: Strategy
+
+    # Data
     measurements_exp: pd.DataFrame = Field(default_factory=pd.DataFrame)
+    numerical_measurements_must_be_within_tolerance: bool = True
+
+    # Metadata
     batches_done: int = 0
     fits_done: int = 0
 
     # TODO: make private
     cached_recommendation: pd.DataFrame = Field(default_factory=pd.DataFrame)
-
-    @property
-    def searchspace(self) -> SearchSpace:
-        """The underlying search space."""
-        return self.strategy.searchspace
 
     @property
     def parameters(self) -> List[Parameter]:
@@ -81,7 +83,6 @@ class BayBE(BaseModel):
             for k in [
                 "allow_repeated_recommendations",
                 "allow_recommending_already_measured",
-                "numerical_measurements_must_be_within_tolerance",
             ]
             if k in config
         }
@@ -96,10 +97,10 @@ class BayBE(BaseModel):
         # parsing process to call Searchspace.create instead of the default constructor
         parameters = parse_obj_as(List[Parameter], config["parameters"])
         if "constraints" in config:
-            constraints = parse_obj_as(List[Constraint], config["constraints"])
+            constraints = parse_obj_as(List[Constraint], config.pop("constraints"))
         else:
             constraints = None
-        config["strategy"]["searchspace"] = SearchSpace.create(
+        config["searchspace"] = SearchSpace.create(
             parameters=parameters,
             constraints=constraints,
         )
@@ -157,7 +158,7 @@ class BayBE(BaseModel):
         # Update meta data
         # TODO: refactor responsibilities
         self.searchspace.discrete.mark_as_measured(
-            data, self.strategy.numerical_measurements_must_be_within_tolerance
+            data, self.numerical_measurements_must_be_within_tolerance
         )
 
         # Read in measurements and add them to the database
@@ -206,9 +207,10 @@ class BayBE(BaseModel):
 
         # Get the recommended search space entries
         rec = self.strategy.recommend(
+            self.searchspace,
+            batch_quantity,
             self.measurements_parameters_comp,
             self.measurements_targets_comp,
-            batch_quantity,
         )
 
         # Query user input